Enum and other minor modernisations

+++ 3escore/PacketHeader.h
@@ -27,7 +27,7 @@ extern const uint16_t TES_CORE_API kPacketCompatibilityVersionMajor;
 extern const uint16_t TES_CORE_API kPacketCompatibilityVersionMinor;
 
 /// Flag values for @c PacketHeader objects.
-enum PacketFlag : unsigned
+enum PacketFlag : uint8_t
 {
   /// Marks a @c PacketHeader as missing its 16-bit CRC.
   PFNoCrc = (1u << 0u),
